DRTII-882 simplity Api file by separate code to specific class
